backend/requirements.txt
Flask==2.1.3
Flask_Caching==2.1.0
Flask_Cors==4.0.0
Flask_Login==0.6.2
Flask_RESTful==0.3.10
Flask_SQLAlchemy==2.5.1
PyJWT==1.7.1
SQLAlchemy==1.4.49
Werkzeug==2.3.7
gunicorn
celery==5.2.3
importlib-metadata<5.0
psycopg2-binary==2.9.9
redis
matplotlib